Index and standardize all redaction syntaxes
**Before**, when redacting, we would redact different kinds of source syntax
to different kinds of redacted syntax.

For example, links like `[a link](http://link-url.com)` would be
redacted to `[a link]`, and images like `![an
image](http://image.com/url.jpg)` would be redacted to `![an image]`.

This means that the redacted syntax received two bits of information;
data about the _type_ of redaction (link or image) and some content to
be translated ("a link" or "an image").

The presence of the type data implied that we would want to continue to
create new, unique syntaxes for all the new kinds of things we intend to
eventually redact, which is a relatively unsustainable process.
Additionally, the _absence_ of any kind of ordering or index data means
that translations that result in a reordering of the redacted values
(due to sentence restructuring, for example) would end up getting messed
up.

**Now**, when redacting, we would instead like to redact everything to
the same syntax, and maintain an ordering. So links like `[a
link](http://link-url.com)` will be redacted to `[a link][0]` and images
like `![an image](http://image.com/url.jpg)` would be redacted to `[an
image][0]`. A bit of content that contained a link followed by an image
would result in them being indexed like `[a link][0]` and `[an
image][1]`