Don't add newlines for serialized data
For context, remark-stringify [adds a newline to the end of all rendered
strings to ensure that they are valid POSIX
files](https://github.com/remarkjs/remark/blob/remark-stringify%405.0.0/packages/remark-stringify/lib/visitors/root.js#L6).
This is all well and good, except that when we are processing serialized
data (as in a JSON or YAML object, which most of our i18n data is), we
don't care about the individual strings within that data being valid
POSIX files since they're not files at all.

The simple fix is to strip off the newline when our recursive string
processors notice that they are processing an object.