Allow for processing intersections of data
When recursively iterating over pairs of input (which we do in the
restoration process), we check against two things: that all given inputs
are defined, and that they are of the same type.

Unfortunately, we're actually redundantly checking the latter before we
check the former, which means that we can't (as intended) deal with a
situation where the redacted data is a subset of the source data - which
is precisely the situation we are in when not all data has been
translated.

The simple fix is to swap the order of the checks, so intersections work
as intended.