Fix locator for redaction restoration
The "locator" method for inline tokenizers is supposed to, given some
string, indicate the next position in the string at which the relevant
plugin might be able to begin tokenizing. So, for example, the redaction
restoration localizer should look at:

    change [COLOR][0] [METHOD][1] by [AMOUNT][2]

And return `7`. Unfortunately, the locator method was actually using the
same regex that we use in the actual tokenizer itself, which includes
(by design) a "at beginning of string" check, meaning that for the above
string the locator was returning `-1`. (As an aside, this is why the
restoration was still working for strings like `[GROUP][0] do [MOVE][1]
[DIR][2] forever`, because the redaction was at the beginning).

**However!** This error was being masked by the fact that the locators
for links and references _also_ just check for an open square bracket;
the locators are used by the parser just to decide when to stop and
check for a matching tokenizer, not to determine _which_ tokenizers to
check. So links and references were telling the parser to stop at `7`,
at which point it found the redaction it would not otherwise have tried
to look for.

This bug only manifested in the text parser, which removes the link and
reference plugins.

tl;dr: the locator logic for the redaction restoration process has never been
accurate, but that flaw has been masked by the presence of the locators
for links and references.